Fixed ff_biotac hand description, merged new tests into a single file
