+++ sr_description/robots/shadowhand_motor_ff_biotac.urdf.xacro
@@ -5,12 +5,9 @@
        name="shadowhand_motor">
 
   <xacro:include filename="$(find sr_description)/materials.urdf.xacro" />
-  <xacro:include filename="$(find sr_description)/hand/xacro/full_hand_motor_ff_biotac.urdf.xacro" />
+  <xacro:include filename="$(find sr_description)/hand/xacro/ff_biotac_hand.urdf.xacro" />
   <xacro:include filename="$(find sr_description)/other/xacro/gazebo/gazebo.urdf.xacro" />
 
-<!-- C6 hand with BioTAC sensor
-  (C) 2012 fnh, hendrich@informatik.uni-hamburg.de
--->
 	<!-- bio is used for all other fingers since ff is forced at bio -->
   <xacro:shadowhand muscletrans="false" muscle="false" bio="false" ubi="false" eli="false" reflect="1.0" prefix="" />
