--- sr_description/test/test_urdf.test
@@ -1,3 +1,0 @@
-<launch>
-  <test test-name="test_urdf" pkg="sr_description" type="test_urdf" time-limit="180.0"/>
-</launch>
